+++ packages/aws-client/S3.js
@@ -4,7 +4,6 @@ const path = require('path');
 const pMap = require('p-map');
 const pRetry = require('p-retry');
 const pump = require('pump');
-const range = require('lodash/range');
 const url = require('url');
 
 const {
@@ -17,6 +16,7 @@ const {
 } = require('@cumulus/errors');
 const Logger = require('@cumulus/logger');
 
+const S3MultipartUploads = require('./lib/S3MultipartUploads');
 const awsServices = require('./services');
 const { inTestMode } = require('./test-utils');
 const { improveStackTrace } = require('./utils');
@@ -623,111 +623,6 @@ exports.getFileBucketAndKey = (pathParams) => {
 exports.createBucket = (Bucket) =>
   awsServices.s3().createBucket({ Bucket }).promise();
 
-const GB = 1024 * 1024 * 1024;
-
-// Each part of a multi-part copy needs to specify a byte range to be copied.
-// This byte range has a starting byte and an ending byte (inclusive) that makes
-// up the part. All parts must be larger than 5 MB, except that the last part
-// may be smaller than 5 MB. Parts cannot be larger than 5 GB.
-//
-// This function takes a file size and an optional maxSize. It returns an array
-// of objects, each containing a `start` and an `end` value. These will make up
-// the ranges of the multi-part copy.
-//
-// https://docs.aws.amazon.com/AmazonS3/latest/API/API_UploadPartCopy.html
-const createMultipartChunks = (size, maxSize = 5 * GB) => {
-  const lastChunkSize = size % maxSize;
-
-  // Build the list of full-size chunks
-  const chunks = range(0, size - lastChunkSize, maxSize).map((start) => {
-    const end = start + maxSize - 1;
-    return { start, end };
-  });
-
-  // If necessary build the last, not-full-size chunk
-  if (lastChunkSize !== 0) {
-    const start = size - lastChunkSize;
-    const end = size - 1;
-    chunks.push({ start, end });
-  }
-
-  return chunks;
-};
-// exported to support testing
-exports.createMultipartChunks = createMultipartChunks;
-
-const createMultipartUpload = async (params) => {
-  const response = await awsServices.s3().createMultipartUpload(params).promise();
-  return response.UploadId;
-};
-
-const completeMultipartUpload = async (params) => {
-  await awsServices.s3().completeMultipartUpload(params).promise();
-};
-
-const abortMultipartUpload = (params) =>
-  awsServices.s3().abortMultipartUpload(params).promise();
-
-// This performs an S3 `uploadPartCopy` call. That response includes an `ETag`
-// value specific to the part that was uploaded. When `completeMultipartUpload`
-// is called later, it needs that `ETag` value, as well as the `PartNumber` for
-// each part. Since the `PartNumber` is not included in the `uploadPartCopy`
-// response, we are adding it here to make our lives easier when we eventually
-// call `completeMultipartUpload`.
-const uploadPartCopy = async (params) => {
-  const response = await awsServices.s3().uploadPartCopy(params).promise();
-
-  return {
-    ...response,
-    PartNumber: params.PartNumber
-  };
-};
-
-// Given a list of chunks (as generated by `createMultipartChunks`), create the
-// list of params that will be passed to each invocation of the `uploadPartCopy`
-// function.
-const buildUploadPartCopyParams = ({
-  chunks,
-  destinationBucket,
-  destinationKey,
-  sourceBucket,
-  sourceKey,
-  uploadId
-}) =>
-  chunks.map(({ start, end }, index) => ({
-    UploadId: uploadId,
-    Bucket: destinationBucket,
-    Key: destinationKey,
-    PartNumber: index + 1,
-    CopySource: `/${sourceBucket}/${sourceKey}`,
-    CopySourceRange: `bytes=${start}-${end}`
-  }));
-// exported to support testing
-exports.buildUploadPartCopyParams = buildUploadPartCopyParams;
-
-// Generate the params that will be passed to the `completeMultipartUpload`
-// function
-const buildCompleteMultipartUploadParams = ({
-  destinationBucket,
-  destinationKey,
-  uploadId,
-  uploadPartCopyResponses
-}) => ({
-  UploadId: uploadId,
-  Bucket: destinationBucket,
-  Key: destinationKey,
-  MultipartUpload: {
-    Parts: uploadPartCopyResponses.map(
-      (response) => ({
-        PartNumber: response.PartNumber,
-        ETag: response.CopyPartResult.ETag
-      })
-    )
-  }
-});
-// exported to support testing
-exports.buildCompleteMultipartUploadParams = buildCompleteMultipartUploadParams;
-
 /**
  * Copy an S3 object to another location in S3 using a multipart copy
  *
@@ -736,8 +631,11 @@ exports.buildCompleteMultipartUploadParams = buildCompleteMultipartUploadParams;
  * @param {string} params.sourceKey
  * @param {string} params.destinationBucket
  * @param {string} params.destinationKey
+ * @returns {Promise<undefined>}
  */
-const multipartCopyObject = async (params = {}) => {
+const multipartCopyObject = async (params) => {
+  debugger;
+
   const {
     sourceBucket,
     sourceKey,
@@ -745,17 +643,19 @@ const multipartCopyObject = async (params = {}) => {
     destinationKey
   } = params;
 
-  const uploadId = await createMultipartUpload({
+  // Create a multi-part upload (copy) and get its UploadId
+  const uploadId = await S3MultipartUploads.createMultipartUpload({
     Bucket: destinationBucket,
     Key: destinationKey
   });
 
   try {
+    // Build the separate parts of the multi-part upload (copy)
     const objectSize = await exports.getObjectSize(sourceBucket, sourceKey);
 
-    const chunks = createMultipartChunks(objectSize);
+    const chunks = S3MultipartUploads.createMultipartChunks(objectSize);
 
-    const uploadPartCopyParams = buildUploadPartCopyParams({
+    const uploadPartCopyParams = S3MultipartUploads.buildUploadPartCopyParams({
       chunks,
       destinationBucket,
       destinationKey,
@@ -764,20 +664,24 @@ const multipartCopyObject = async (params = {}) => {
       uploadId
     });
 
+    // Submit all of the upload (copy) parts to S3
     const uploadPartCopyResponses = await Promise.all(
-      uploadPartCopyParams.map(uploadPartCopy)
+      uploadPartCopyParams.map((p) => S3MultipartUploads.uploadPartCopy(p))
     );
 
-    const completeMultipartUploadParams = buildCompleteMultipartUploadParams({
+    // Let S3 know that the multi-part upload (copy) is completed
+    const completeMultipartUploadParams = S3MultipartUploads.buildCompleteMultipartUploadParams({
       uploadPartCopyResponses,
       destinationBucket,
       destinationKey,
       uploadId
     });
 
-    await completeMultipartUpload(completeMultipartUploadParams);
+    await S3MultipartUploads.completeMultipartUpload(completeMultipartUploadParams);
   } catch (error) {
-    await abortMultipartUpload({
+    // If anything went wrong, make sure that the multi-part upload (copy)
+    // is aborted.
+    await S3MultipartUploads.abortMultipartUpload({
       Bucket: destinationBucket,
       Key: destinationKey,
       UploadId: uploadId
