+++ packages/aws-client/tests/S3/test-multipartCopyObjectFailure.js
@@ -6,31 +6,23 @@ const proxyquire = require('proxyquire');
 let abortMultipartUploadWasCalled = false;
 let abortMultipartUploadParams;
 
-const fakeS3Service = () => ({
-  abortMultipartUpload: (params) => ({
-    promise: async () => {
-      abortMultipartUploadWasCalled = true;
-      abortMultipartUploadParams = params;
-    }
-  }),
-  createMultipartUpload: () => ({
-    promise: () => Promise.resolve({
-      UploadId: 'abc-123'
-    })
-  }),
-  headObject: () => ({
-    promise: () => Promise.resolve({ ContentLength: 5 })
-  }),
-  uploadPartCopy: () => ({
-    promise: () => Promise.reject(new Error('uh oh'))
-  })
-});
-
 const S3 = proxyquire(
   '../../S3',
   {
     './services': {
-      s3: fakeS3Service
+      s3: () => ({
+        headObject: () => ({
+          promise: () => Promise.resolve({ ContentLength: 5 })
+        })
+      })
+    },
+    './lib/S3MultipartUploads': {
+      abortMultipartUpload: async (params) => {
+        abortMultipartUploadWasCalled = true;
+        abortMultipartUploadParams = params;
+      },
+      createMultipartUpload: () => Promise.resolve('abc-123'),
+      uploadPartCopy: () => Promise.reject(new Error('uh oh'))
     }
   }
 );
